Batch rename files to .jsx
