--- src/components/LanguageSwitcher.js
@@ -1,6 +1,0 @@
-import React from 'react';
-import Typography from '@material-ui/core/Typography'
-
-export default function LanguageSwitcher(props) {
-    return <Typography>English</Typography>
-}
