+++ src/components/LanguageSwitcher.jsx
@@ -1,0 +1,6 @@
+import React from 'react';
+import Typography from '@material-ui/core/Typography';
+
+export default function LanguageSwitcher(props) {
+    return <Typography>English</Typography>;
+}
